% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASAP_SIAM_to_ScantronAS.R
\name{ASAP.SIAM_to_ScantronAS}
\alias{ASAP.SIAM_to_ScantronAS}
\title{ASAP Student Item Analysis Multi to Scantron Achievement Series}
\usage{
ASAP.SIAM_to_ScantronAS(asapFilePath, sectionTable, merge = "TP",
  alphaResponses = T, additionalExams = NULL, messageLevel = 0)
}
\arguments{
\item{asapFilePath}{Character of length 1 containing the filepath of an
export from ASAP.}

\item{sectionTable}{data.frame showing associations between students and
sections.  Columns include \code{StudentID}, \code{Course}, \code{Teacher},
\code{Period}, and \code{Exam}. Each Exam value must match to exactly 1
regents exam name using \code{grepl}.}

\item{merge}{character of length 1 indicating when to merge sections.}

\item{alphaResponses}{logical of length 1 - should MC responses be converted
from digits to letters?  Defaults to TRUE.}

\item{messageLevel}{integer of length 1 indicating the level of messaging to
print.  Defaults to \code{TP}.}
}
\value{
Nothing is returned by this function.
}
\description{
Convert the Student Item Analysis Multi export from Level 0 ASAP
  to the format of files exported from Scantron Achievement Series
}
\details{
For the \code{merge} parameter, the value will indicate whether to
  group different sections together.  The most grouped values are \code{T},
  \code{C}, or \code{P} (for grouping all students who have the same teacher,
  course, or period, respectively).  The least grouped value is \code{TCP}
  (for grouping students only when they have the exact same teacher, course,
  and period).  Other acceptable values are \code{TC}, \code{TP}, and
  \code{CP}.  Defaults to \code{TP}.

  The export files generated by this function will be placed in a folder
  named \code{export} within the folder containing \code{asapFilePath}.
}
